Tests/LibWeb/Text/input/Fetch/fetch-later-activate-after-zero.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script src="fetch-later-helper.js"></script>
<script>
    promiseTest(async () => {
        try {
            let server = httpTestServer();
            let token = `${Date.now()}-${Math.random().toString(16).slice(2)}`;
            let cookieName = `fetchlater_zero_${token.replace("-", "_")}`;

            let checkUrl = await server.createEcho("GET", `/fetch-later-zero-check-${token}`, {
                status: 200,
                headers: {
                    "Access-Control-Allow-Origin": location.origin,
                    "Access-Control-Allow-Credentials": "true",
                },
                reflect_headers_in_body: true,
            });
            let setUrl = await server.createEcho("POST", `/fetch-later-zero-set-${token}`, {
                status: 204,
                headers: {
                    "Access-Control-Allow-Origin": location.origin,
                    "Access-Control-Allow-Credentials": "true",
                    "Set-Cookie": `${cookieName}=1; Path=/`,
                },
            });

            let result = fetchLater(setUrl, {
                method: "POST",
                credentials: "include",
                body: "activate now",
                activateAfter: 0,
            });

            println(`activated before wait: ${result.activated}`);
            let activated = await waitForActivated(result);
            let cookieObserved = await waitForCookie(checkUrl, cookieName);
            println(`activated after wait: ${activated}`);
            println(`cookie observed: ${cookieObserved}`);
            println(activated && cookieObserved ? "PASS" : "FAIL");
        } catch (error) {
            println(`FAIL - ${error}`);
        }
    });
</script>
