Tests/LibWeb/Text/expected/Fetch/fetch-later-activate-after-zero.txt

activated before wait: false
activated after wait: true
cookie observed: true
PASS
